+++ collector/conntrack_linux.go
@@ -16,10 +16,6 @@
 package collector
 
 import (
-	"io/ioutil"
-	"strconv"
-	"strings"
-
 	"github.com/prometheus/client_golang/prometheus"
 )
 
@@ -66,16 +62,4 @@ func (c *conntrackCollector) Update(ch chan<- prometheus.Metric) (err error) {
 		c.limit, prometheus.GaugeValue, float64(value))
 
 	return nil
-}
-
-func readUintFromFile(path string) (uint64, error) {
-	data, err := ioutil.ReadFile(path)
-	if err != nil {
-		return 0, err
-	}
-	value, err := strconv.ParseUint(strings.TrimSpace(string(data)), 10, 64)
-	if err != nil {
-		return 0, err
-	}
-	return value, nil
 }
